+++ csharp_pathfinding/A_star/Solvers/IPathfindingSolver.cs
@@ -5,7 +5,7 @@ namespace AStarNickNS
 {
     public interface IPathfindingSolver<out TPlace, TCoord> where TPlace : IPlace<TCoord>
     {
-        public IEnumerable<TPlace> SolvePath(IPlace<TCoord> start, IPlace<TCoord> target, CancellationToken token, double pathfinderSize);
+        public IEnumerable<TPlace> SolvePath(IPlace<TCoord> start, IPlace<TCoord> target, CancellationToken token, float pathfinderSize);
 
     }
 }
